added puppeteer & improved test scripts

+++ scripts/e2eCleanup.js
@@ -12,6 +12,7 @@ if (existsSync(testDbPath)) {
 if (existsSync(testDbJournalPath)) {
     unlinkSync(testDbJournalPath);
 }
-
+if (existsSync(tempPrismaEnvPath)) {
+    renameSync(tempPrismaEnvPath, prismaEnvPath);
+}
 unlinkSync(prismaEnvPath);
-renameSync(tempPrismaEnvPath, prismaEnvPath);
